classicCount-class {topGO}	R Documentation

Class "classicCount"

Description

This class that extends the virtual class "groupStats" by adding a slot representing the significant members.

Details

This class is used for test statistic based on counts, like Fisher's exact test

Objects from the Class

Objects can be created by calls of the form new("classicCount", testStatistic = "function", name = "character", allMembers = "character", groupMembers = "character", sigMembers = "character").

Slots

significant:

Object of class "integer" ~~

name:

Object of class "character" ~~

allMembers:

Object of class "character" ~~

members:

Object of class "character" ~~

testStatistic:

Object of class "function" ~~

Extends

Class "groupStats", directly.

Methods

contTable

signature(object = "classicCount"): ...

initialize

signature(.Object = "classicCount"): ...

numSigAll

signature(object = "classicCount"): ...

numSigMembers

signature(object = "classicCount"): ...

sigAllMembers

signature(object = "classicCount"): ...

sigMembers<-

signature(object = "classicCount"): ...

sigMembers

signature(object = "classicCount"): ...

Author(s)

Adrian Alexa

See Also

classicScore-class, groupStats-class, getSigGroups-methods

Examples

##---- Should be DIRECTLY executable !! ----

[Package topGO version 2.34.0 Index]