Class "classicScore"

A class that extends the virtual class "groupStats" by adding a slot representing the score of each gene. It is used for tests like Kolmogorov-Smirnov test.

Objects from the Class

Objects can be created by calls of the form new("classicScore", testStatistic, name, allMembers, groupMembers, score, decreasing).

Slots

score:

Object of class "numeric" ~~

name:

Object of class "character" ~~

allMembers:

Object of class "character" ~~

members:

Object of class "character" ~~

testStatistic:

Object of class "function" ~~

scoreOrder:

Object of class "character" ~~

testStatPar:

Object of class "ANY" ~~

Extends

Class "groupStats", directly.

Methods

allScore

Method to obtain the score of all members.

scoreOrder

Returns TRUE if the score should be ordered increasing, FALSE otherwise.

membersScore

signature(object = "classicScore"): ...

rankMembers

signature(object = "classicScore"): ...

score<-

signature(object = "classicScore"): ...

Author

Adrian Alexa

See also

Examples

## define the type of test you want to use
test.stat <- new("classicScore", testStatistic = GOKSTest, name = "KS tests")