elimExpr-class {topGO}	R Documentation

Class "elimExpr"

Description

Classes that extend the "classicExpr" class by adding a slot representing the members that need to be removed.

Details

TODO: Some datails here.....

Objects from the Class

Objects can be created by calls of the form new("elimExpr", testStatistic, name, groupMembers, exprDat, pType, elim, cutOff, ...). ~~ describe objects here ~~

Slots

cutOff:

Object of class "numeric" ~~

elim:

Object of class "integer" ~~

eData:

Object of class "environment" ~~

pType:

Object of class "factor" ~~

name:

Object of class "character" ~~

allMembers:

Object of class "character" ~~

members:

Object of class "character" ~~

testStatistic:

Object of class "function" ~~

testStatPar:

Object of class "list" ~~

Extends

Class "weight01Expr", directly. Class "classicExpr", by class "weight01Expr", distance 2. Class "groupStats", by class "weight01Expr", distance 3.

Methods

cutOff<-

signature(object = "elimExpr"): ...

cutOff

signature(object = "elimExpr"): ...

getSigGroups

signature(object = "topGOdata", test.stat = "elimExpr"): ...

initialize

signature(.Object = "elimExpr"): ...

Author(s)

Adrian Alexa

See Also

classicScore-class, groupStats-class, getSigGroups-methods

Examples

showClass("elimExpr")

[Package topGO version 2.34.0 Index]