A toy example of a list of gene identifiers and the respective p-values

The geneList data is compiled from a differential expression analysis of the ALL dataset. It contains just a small number of genes with the corespondent p-values. The information on where to find the GO annotations is stored in the ALL object.

The topDiffGenes function included in this dataset will select the differentially expressed genes, at 0.01 significance level, from geneList.

Usage

data(geneList)

Source

Generated using the ALL gene expression data. See the "scripts" directory.

Examples

data(geneList)

## print the object
head(geneList)
#> 1095_s_at   1130_at   1196_at 1329_s_at 1340_s_at 1342_g_at 
#> 1.0000000 1.0000000 0.6223795 0.5412240 1.0000000 1.0000000 
length(geneList)
#> [1] 323

## the number of genes with a p-value less than 0.01
sum(topDiffGenes(geneList))
#> [1] 50