groupGOTerms {topGO}	R Documentation

Grouping of GO terms into the three ontologies

Description

This function split the GOTERM environment into three different ontologies. The newly created environments contain each only the terms from one of the following ontologies 'BP', 'CC', 'MF'

Usage

groupGOTerms(where)

Arguments

where
The the environment where you want to bind the results.

Value

The function returns NULL.

Author(s)

Adrian Alexa

See Also

topGOdata-class, GOTERM

Examples

groupGOTerms()

[Package topGO version 2.34.0 Index]