groupStats-class {topGO}	R Documentation

Class "groupStats"

Description

A virtual class containing basic gene set information: the gene universe, the member of the current group, the test statistic defined for this group, etc.

Objects from the Class

A virtual Class: No objects may be created from it.

Slots

name:

Object of class "character" ~~

allMembers:

Object of class "character" ~~

members:

Object of class "character" ~~

testStatistic:

Object of class "function" ~~

testStatPar:

Object of class "ANY" ~~

Methods

allMembers<-

signature(object = "groupStats"): ...

allMembers

signature(object = "groupStats"): ...

initialize

signature(.Object = "groupStats"): ...

members<-

signature(object = "groupStats"): ...

members

signature(object = "groupStats"): ...

Name<-

signature(object = "groupStats"): ...

Name

signature(object = "groupStats"): ...

numAllMembers

signature(object = "groupStats"): ...

numMembers

signature(object = "groupStats"): ...

runTest

signature(object = "groupStats"): ...

testStatistic

signature(object = "groupStats"): ...

Author(s)

Adrian Alexa

See Also

classicCount-class, getSigGroups-methods

[Package topGO version 2.34.0 Index]