classicExpr-class {topGO}	R Documentation

Class "classicExpr"

Description

This class that extends the virtual class "groupStats" by adding two slots for accomodating gene expression data.

Objects from the Class

Objects can be created by calls of the form new("classicExpr", testStatistic, name, groupMembers, exprDat, pType, ...).

Slots

eData:

Object of class "environment" ~~

pType:

Object of class "factor" ~~

name:

Object of class "character" ~~

allMembers:

Object of class "character" ~~

members:

Object of class "character" ~~

testStatistic:

Object of class "function" ~~

testStatPar:

Object of class "list" ~~

Extends

Class "groupStats", directly.

Methods

allMembers<-

signature(object = "classicExpr"): ...

emptyExpr

signature(object = "classicExpr"): ...

getSigGroups

signature(object = "topGOdata", test.stat = "classicExpr"): ...

GOglobalTest

signature(object = "classicExpr"): ...

initialize

signature(.Object = "classicExpr"): ...

membersExpr

signature(object = "classicExpr"): ...

pType<-

signature(object = "classicExpr"): ...

pType

signature(object = "classicExpr"): ...

Author(s)

Adrian Alexa

See Also

classicScore-class, groupStats-class, getSigGroups-methods

Examples

showClass("classicExpr")

[Package topGO version 2.34.0 Index]