Class "classicExpr"

This class that extends the virtual class "groupStats" by adding two slots for accommodating gene expression data.

Objects from the Class

Objects can be created by calls of the form new("classicExpr", testStatistic, name, groupMembers, exprDat, pType, ...).

Slots

eData:

Object of class "environment" ~~

pType:

Object of class "factor" ~~

name:

Object of class "character" ~~

allMembers:

Object of class "character" ~~

members:

Object of class "character" ~~

testStatistic:

Object of class "function" ~~

testStatPar:

Object of class "list" ~~

Extends

Class "groupStats", directly.

Methods

allMembers<-

signature(object = "classicExpr"): ...

emptyExpr

signature(object = "classicExpr"): ...

getSigGroups

signature(object = "topGOdata", test.stat = "classicExpr"): ...

GOglobalTest

signature(object = "classicExpr"): ...

initialize

signature(.Object = "classicExpr"): ...

membersExpr

signature(object = "classicExpr"): ...

pType<-

signature(object = "classicExpr"): ...

pType

signature(object = "classicExpr"): ...

Author

Adrian Alexa

See also

Examples

showClass("classicExpr")
#> Class "classicExpr" [package "topGO"]
#> 
#> Slots:
#>                                                                             
#> Name:          eData         pType          name    allMembers       members
#> Class:   environment        factor     character     character     character
#>                                   
#> Name:  testStatistic   testStatPar
#> Class:      function          list
#> 
#> Extends: "groupStats"
#> 
#> Known Subclasses: 
#> Class "weight01Expr", directly
#> Class "elimExpr", by class "weight01Expr", distance 2
#> Class "leaExpr", by class "weight01Expr", distance 2