Interfaces for running the enrichment tests

These function are used for dispatching the specific algorithm for a given topGOdata object and a test statistic.

getSigGroups(object, test.stat, ...)
runTest(object, algorithm, statistic, ...)
whichAlgorithms()
whichTests()

Arguments

object

An object of class topGOdata This object contains all data necessary for running the test.

test.stat

An object of class groupStats. This object defines the test statistic.

algorithm

Character string specifying which algorithm to use.

statistic

Character string specifying which test to use.

...

Other parameters. In the case of runTest they are used for defining the test statistic

Details

The runTest function can be used only with a predefined set of test statistics and algorithms. The algorithms and the statistical tests which are accessible via the runTest function are shown by the whichAlgorithms() and whichTests() functions.

The runTest function is a warping of the getSigGroups and the initialisation of a groupStats object functions.

...

Value

An object of class topGOresult.

Author

Adrian Alexa

See also

Examples

## load a sample topGOdata object
data(GOdata)
GOdata
#> 
#> ------------------------- topGOdata object -------------------------
#> 
#>  Description:
#>    -  Simple topGOdata object 
#> 
#>  Ontology:
#>    -  BP 
#> 
#>  323 available genes (all genes from the array):
#>    - symbol:  1095_s_at 1130_at 1196_at 1329_s_at 1340_s_at  ...
#>    - score :  1 1 0.62238 0.541224 1  ...
#>    - 50  significant genes. 
#> 
#>  318 feasible genes (genes that can be used in the analysis):
#>    - symbol:  1095_s_at 1130_at 1196_at 1329_s_at 1340_s_at  ...
#>    - score :  1 1 0.62238 0.541224 1  ...
#>    - 50  significant genes. 
#> 
#>  GO graph (nodes with at least  5  genes):
#>    - a graph with directed edges
#>    - number of nodes = 777 
#>    - number of edges = 1557 
#> 
#> ------------------------- topGOdata object -------------------------
#> 

##############################
## getSigGroups interface
##############################

## define a test statistic
test.stat <- new("classicCount", testStatistic = GOFisherTest, name = "Fisher test")
## perform the test
resultFis <- getSigGroups(GOdata, test.stat)
#> 
#> 			 -- Classic Algorithm -- 
#> 
#> 		 the algorithm is scoring 590 nontrivial nodes
#> 		 parameters: 
#> 			 test statistic: Fisher test
resultFis
#> 
#> Description: Simple topGOdata object 
#> Ontology: BP 
#> 'classic' algorithm with the 'Fisher test' test
#> 777 GO terms scored: 48 terms with p < 0.01
#> Annotation data:
#>     Annotated genes: 318 
#>     Significant genes: 50 
#>     Min. no. of genes annotated to a GO: 5 
#>     Nontrivial nodes: 590 

##############################
## runTest interface
##############################

## Enrichment analysis by using the "classic" method and Fisher's exact test
resultFis <- runTest(GOdata, algorithm = "classic", statistic = "fisher")
#> 
#> 			 -- Classic Algorithm -- 
#> 
#> 		 the algorithm is scoring 590 nontrivial nodes
#> 		 parameters: 
#> 			 test statistic: fisher
resultFis
#> 
#> Description: Simple topGOdata object 
#> Ontology: BP 
#> 'classic' algorithm with the 'fisher' test
#> 777 GO terms scored: 48 terms with p < 0.01
#> Annotation data:
#>     Annotated genes: 318 
#>     Significant genes: 50 
#>     Min. no. of genes annotated to a GO: 5 
#>     Nontrivial nodes: 590 

## weight01 is the default algorithm 
weight01.fisher <- runTest(GOdata, statistic = "fisher")
#> 
#> 			 -- Weight01 Algorithm -- 
#> 
#> 		 the algorithm is scoring 590 nontrivial nodes
#> 		 parameters: 
#> 			 test statistic: fisher
#> 
#> 	 Level 13:	2 nodes to be scored	(0 eliminated genes)
#> 
#> 	 Level 12:	6 nodes to be scored	(0 eliminated genes)
#> 
#> 	 Level 11:	15 nodes to be scored	(12 eliminated genes)
#> 
#> 	 Level 10:	29 nodes to be scored	(41 eliminated genes)
#> 
#> 	 Level 9:	42 nodes to be scored	(76 eliminated genes)
#> 
#> 	 Level 8:	62 nodes to be scored	(147 eliminated genes)
#> 
#> 	 Level 7:	99 nodes to be scored	(222 eliminated genes)
#> 
#> 	 Level 6:	108 nodes to be scored	(305 eliminated genes)
#> 
#> 	 Level 5:	98 nodes to be scored	(313 eliminated genes)
#> 
#> 	 Level 4:	72 nodes to be scored	(315 eliminated genes)
#> 
#> 	 Level 3:	41 nodes to be scored	(315 eliminated genes)
#> 
#> 	 Level 2:	15 nodes to be scored	(316 eliminated genes)
#> 
#> 	 Level 1:	1 nodes to be scored	(316 eliminated genes)
weight01.fisher
#> 
#> Description: Simple topGOdata object 
#> Ontology: BP 
#> 'weight01' algorithm with the 'fisher' test
#> 777 GO terms scored: 11 terms with p < 0.01
#> Annotation data:
#>     Annotated genes: 318 
#>     Significant genes: 50 
#>     Min. no. of genes annotated to a GO: 5 
#>     Nontrivial nodes: 590 

## not all combinations are possible!
# weight.ks <- runTest(GOdata, algorithm = "weight", statistic = "t")