Visualisation functions

Functions to plot the subgraphs induced by the most significant GO terms

showSigOfNodes(GOdata, termsP.value, firstSigNodes = 10, reverse = TRUE,
               sigForAll = TRUE, wantedNodes = NULL, putWN = TRUE,
               putCL = 0, type = NULL, showEdges = TRUE,  swPlot = TRUE,
               useFullNames = TRUE, oldSigNodes = NULL, 
               useInfo = c("none", "pval", "counts", "def", "np", "all")[1],
               plotFunction = GOplot, .NO.CHAR = 20)

printGraph(object, result, firstSigNodes, refResult, ...)

Arguments

object

an object of class topGOdata.

GOdata

an object of class topGOdata.

result

an object of class topGOresult.

firstSigNodes

the number of top scoring GO terms which ....

refResult

an object of class topGOresult.

termsP.value

named vector of p-values.

reverse

the direction of the edges.

sigForAll

if TRUE the score/p-value of all nodes in the DAG is shown, otherwise only the score for the sigNodes

wantedNodes

the nodes that we want to find, we will plot this nodes with a different color. The vector contains the names of the nodes

putWN

the graph is generated with using the firstSigNodes and the wantedNodes.

putCL

we generate the graph from the nodes given by all previous parameters, plus their children. if putCL = 1 than only the children are added, if putCL = n we get the nodes form the next n levels.

type

used for plotting pie charts

showEdges

if TRUE the edge are shown

swPlot

if true the graph is ploted, if not no plotting is done.

useInfo

additional info to be ploted to each node.

oldSigNodes

used to plot the (new) sigNodes in the same color range as the old ones

useFullNames

argument for internal use ..

plotFunction

argument for internal use ..

.NO.CHAR

argument for internal use ..

...

Extra arguments for printGraph can be:

fn.prefix

character string giving the file name prefix.

useInfo

as in showSigOfNodes function.

pdfSW

logical attribute switch between PDF or PS formats.

Details

There are two functions available. The showSigOfNodes will plot the induced subgraph to the current graphic device. The printGraph is a warping function for showSigOfNodes and will save the resulting graph into a PDF or PS file.

In the plots, the significant nodes are represented as rectangles. The plotted graph is the upper induced graph generated by these significant nodes.

Author

Adrian Alexa

See also

Examples

if (FALSE) { # \dontrun{
data(GOdata)
data(results.tGO)

showSigOfNodes(GOdata, score(resultFisher), firstSigNodes = 5, useInfo = 'all')
printGraph(GOdata, resultFisher, firstSigNodes = 5, 
           fn.prefix = "sampleFile", useInfo = "all", pdfSW = TRUE)
} # }