topGOresult-class {topGO}	R Documentation

Class "topGOresult"

Description

Class instance created by getSigGroups-methods or by runTest

Objects from the Class

Objects can be created by calls of the form new("topGOresult", description, score, testName, algorithm, geneData).

Slots

description:

character string containing a short description on how the object was build.

score:

named numerical vector containing the p-values or the scores of the tested GO terms.

testName:

character string containing the name of the test statistic used.

algorithm:

character string containing the name of the algorithm used.

geneData:

list containing summary statistics on the genes/gene universe/annotations.

Methods

score:

method to access the score slot.

testName:

method to access the testName slot.

algorithm:

method to access the algorithm slot.

geneData:

method to access the geneData slot.

show:

method to print the object.

combineResults:

method to aggregate two or more topGOresult objects. method = c("gmean", "mean", "median", "min", "max") provides the way the object scores (which most of the time are p-values) are combined.

.

Author(s)

Adrian Alexa

See Also

groupStats-class, getSigGroups-methods

Examples

data(results.tGO)

s <- score(resultFisher)

go <- sort(names(s))
go.sub<- sample(go, 100)
go.mixed <- c(sample(go, 50), sample(ls(GOCCTerm), 20))
go.others <- sample(ls(GOCCTerm), 100)

str(go)
str(go.sub)
str(go.mixed)
str(go.others)

str(score(resultFisher, whichGO = go))
str(score(resultFisher, whichGO = go.sub))
str(score(resultFisher, whichGO = go.mixed))
str(score(resultFisher, whichGO = go.others))

avgResult <- combineResults(resultFisher, resultKS)
avgResult
combineResults(resultFisher, resultKS, method = "min")

[Package topGO version 2.34.0 Index]