Class "topGOresult"

Class instance created by getSigGroups-methods or by runTest

Objects from the Class

Objects can be created by calls of the form new("topGOresult", description, score, testName, algorithm, geneData).

Slots

description:

character string containing a short description on how the object was build.

score:

named numerical vector containing the p-values or the scores of the tested GO terms.

testName:

character string containing the name of the test statistic used.

algorithm:

character string containing the name of the algorithm used.

geneData:

list containing summary statistics on the genes/gene universe/annotations.

Methods

score:

method to access the score slot.

testName:

method to access the testName slot.

algorithm:

method to access the algorithm slot.

geneData:

method to access the geneData slot.

show:

method to print the object.

combineResults:

method to aggregate two or more topGOresult objects. method = c("gmean", "mean", "median", "min", "max") provides the way the object scores (which most of the time are p-values) are combined.

.

Author

Adrian Alexa

See also

Examples

data(results.tGO)

s <- score(resultFisher)

go <- sort(names(s))
go.sub<- sample(go, 100)
go.mixed <- c(sample(go, 50), sample(ls(GOCCTerm), 20))
go.others <- sample(ls(GOCCTerm), 100)

str(go)
#>  chr [1:777] "GO:0000003" "GO:0000018" "GO:0000060" "GO:0000070" ...
str(go.sub)
#>  chr [1:100] "GO:0007010" "GO:0051495" "GO:0048232" "GO:0030335" ...
str(go.mixed)
#>  chr [1:70] "GO:0055086" "GO:0044255" "GO:0031326" "GO:0045582" ...
str(go.others)
#>  chr [1:100] "GO:0032156" "GO:0097457" "GO:0071752" "GO:0098982" ...

str(score(resultFisher, whichGO = go))
#>  Named num [1:777] 0.285 0.645 1 1 0.918 ...
#>  - attr(*, "names")= chr [1:777] "GO:0000003" "GO:0000018" "GO:0000060" "GO:0000070" ...
str(score(resultFisher, whichGO = go.sub))
#> Warning: Not all GOs from the object were retrived.
#>  Named num [1:100] 0.5731 0.0808 1 1 0.3349 ...
#>  - attr(*, "names")= chr [1:100] "GO:0007010" "GO:0051495" "GO:0048232" "GO:0030335" ...
str(score(resultFisher, whichGO = go.mixed))
#> Warning: Not all specified GOs can be found in the object.
#> Warning: Not all GOs from the object were retrived.
#>  Named num [1:50] 1 0.22407 0.00329 0.08075 0.03622 ...
#>  - attr(*, "names")= chr [1:50] "GO:0055086" "GO:0044255" "GO:0031326" "GO:0045582" ...
str(score(resultFisher, whichGO = go.others))
#> Warning: The specified GO terms could not be found in the object... Returning empty vector
#>  num(0) 

avgResult <- combineResults(resultFisher, resultKS)
avgResult
#> 
#> Description: Simple topGOdata object 
#> Ontology: BP
#> gmean(classic, classic) 
#> 'ensemble' algorithm with the 'fisher, ks' test
#> 777 GO terms scored: 43 terms with p < 0.01
#> Annotation data:
#>     Annotated genes: 318 
#>     Significant genes: 50 
#>     Min. no. of genes annotated to a GO: 5 
#>     Nontrivial nodes: 590 
combineResults(resultFisher, resultKS, method = "min")
#> 
#> Description: Simple topGOdata object 
#> Ontology: BP
#> min(classic, classic) 
#> 'ensemble' algorithm with the 'fisher, ks' test
#> 777 GO terms scored: 66 terms with p < 0.01
#> Annotation data:
#>     Annotated genes: 318 
#>     Significant genes: 50 
#>     Min. no. of genes annotated to a GO: 5 
#>     Nontrivial nodes: 590